+++ contracts/lib/UniswapV3Broker.sol
@@ -227,11 +227,11 @@ library UniswapV3Broker {
 
     /// @dev if twapInterval < 10 (should be less than 1 block), return mark price without twap directly,
     ///      as twapInterval is too short and makes getting twap over such a short period meaningless
-    function getSqrtMarkTwapX96(address pool, uint32 twapInterval) internal view returns (uint160) {
-        return getSqrtMarkTwapX96From(pool, 0, twapInterval);
+    function getSqrtMarketTwapX96(address pool, uint32 twapInterval) internal view returns (uint160) {
+        return getSqrtMarketTwapX96From(pool, 0, twapInterval);
     }
 
-    function getSqrtMarkTwapX96From(
+    function getSqrtMarketTwapX96From(
         address pool,
         uint32 secondsAgo,
         uint32 twapInterval